error correction
In flick_fluo_mod there was an error in reading the series. It was
putting the number of the series in place of the zoom and viceversa.
Corrected now.
The channel number in this file is set to 0. I should change the program
so that you can input it.

The EXAMPLE file has been updated to a version that analyses more
series.